Move `FeedImageDataLoaderSpy` into a shared scope to remove duplication
